Is now using the stable release of SDL2.
And due to this it uses MwLibrary2. Some small code changes was done in
order for everything to compile. But some new problem occured with the
sound. The temporary solution was to make the sound incactive.
In order to run on windows the correct dll files must be used, i.e. dll
files for sdl2.